[vcpkg] rename hash to sha
in preparation for adding [types don't know #] style hashing code

[types don't know #]: http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2014/n3980.html
