+++ toolsrc/src/vcpkg/cmakevars.cpp
@@ -1,4 +1,4 @@
-#include <vcpkg/base/hash.h>
+#include <vcpkg/base/digest.h>
 #include <vcpkg/base/optional.h>
 #include <vcpkg/base/span.h>
 #include <vcpkg/base/system.process.h>
